Remove `include/grpcpp/impl/codegen/service_type.h` (#31282)
