+++ include/grpcpp/ext/orca_service.h
@@ -23,8 +23,8 @@
 #include "absl/time/time.h"
 #include "absl/types/optional.h"
 
-#include <grpcpp/impl/codegen/service_type.h>
 #include <grpcpp/impl/codegen/sync.h>
+#include <grpcpp/impl/service_type.h>
 #include <grpcpp/server_builder.h>
 #include <grpcpp/support/server_callback.h>
 #include <grpcpp/support/slice.h>
