+++ src/compiler/cpp_generator.cc
@@ -150,7 +150,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/server_context.h",
-        "grpcpp/impl/codegen/service_type.h",
+        "grpcpp/impl/service_type.h",
         "grpcpp/impl/codegen/status.h",
         "grpcpp/impl/codegen/stub_options.h",
         "grpcpp/impl/codegen/sync_stream.h",
@@ -1659,7 +1659,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/support/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/server_context.h",
-        "grpcpp/impl/codegen/service_type.h",
+        "grpcpp/impl/service_type.h",
         "grpcpp/impl/codegen/sync_stream.h"};
     std::vector<std::string> headers(headers_strs, array_end(headers_strs));
     PrintIncludes(printer.get(), headers, params.use_system_headers,
